chore: move `chain_id` mod to `chain`
